[refactor] windows: Move `{CharStyle, TextLayout}` to `text.rs`
